[🚚change]: 오타검출 함수 유틸폴더로 이동 - getTypoKrChar => compareChar 네이밍 변경
